pyproject.toml
[tool.poetry]
name = "having-fun-in-human-feedback"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
shiny = "^1.2.1"
python-dotenv = "^1.0.1"
psycopg2 = "^2.9.10"
pandas = "^2.2.3"
jinja2 = "^3.1.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
